[ALPS04262738] binder: fix proc->files use-after-free
proc->files cleanup is initiated by binder_vma_close. Therefore
a reference on the binder_proc is not enough to prevent the
files_struct from being released while the binder_proc still has
a reference. This can lead to an attempt to dereference the
stale pointer obtained from proc->files prior to proc->files
cleanup. This has been seen once in task_get_unused_fd_flags()
when __alloc_fd() is called with a stale "files".

The fix is to protect proc->files with a mutex to prevent cleanup
while in use.

Signed-off-by: Todd Kjos <tkjos@google.com>
Cc: stable <stable@vger.kernel.org> # 4.14
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

MTK-Commit-Id: 81de1e1defcea732ca77d9c5fc3b7095f5e55bad

Change-Id: Ied00230bd2c699fb1033b3220ac739e861a9a1bf
CR-Id: ALPS04262738
Feature: [Module]Binder
Signed-off-by: Jian-Min Liu <jian-min.liu@mediatek.com>