[ALPS04262736] binder: fix race between munmap() and direct reclaim
An munmap() on a binder device causes binder_vma_close() to be called
which clears the alloc->vma pointer.

If direct reclaim causes binder_alloc_free_page() to be called, there
is a race where alloc->vma is read into a local vma pointer and then
used later after the mm->mmap_sem is acquired. This can result in
calling zap_page_range() with an invalid vma which manifests as a
use-after-free in zap_page_range().

The fix is to check alloc->vma after acquiring the mmap_sem (which we
were acquiring anyway) and bail out of binder_alloc_free_page() if it
has changed to NULL.

MTK-Commit-Id: b0a0c2d3f0736e4e2410def5265b6939807535f6

Change-Id: I9d5f80ed890eafc0ccf1f1d77486919a330c0f82
CR-Id: ALPS04262736
Feature: [Module]Binder
Signed-off-by: Frankie Chang <Frankie.Chang@mediatek.com>