[ALPS04823032] ext4: zero out the unused memory region in the extent tree block
This commit zeroes out the unused memory region in the buffer_head
corresponding to the extent metablock after writing the extent header
and the corresponding extent node entries.

This is done to prevent random uninitialized data from getting into
the filesystem when the extent block is synced.

This fixes CVE-2019-11833.

MTK-Commit-Id: 6a9b106cc4eaaeec0d1dd8b16488f14e4989fe1c

Change-Id: If3427e7a675aa223cd5b9307779940e61cb95ead
Signed-off-by: Sriram Rajagopalan <sriramr@arista.com>
Signed-off-by: Theodore Tso <tytso@mit.edu>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Light Hsieh <light.hsieh@mediatek.com>
CR-Id: ALPS04823032
Feature: [Android Default] EXT4 File System