[ALPS04825819] BACKPORT: binder: Set end of SG buffer area properly.
In case the target node requests a security context, the
extra_buffers_size is increased with the size of the security context.
But, that size is not available for use by regular scatter-gather
buffers; make sure the ending of that buffer is marked correctly.

Bug: 136210786
Acked-by: Todd Kjos <tkjos@google.com>
Fixes: ec74136ded79 ("binder: create node flag to request senders security context")
Signed-off-by: Martijn Coenen <maco@android.com>
Cc: stable@vger.kernel.org # 5.1+
Link: https://lore.kernel.org/r/20190709110923.220736-1-maco@android.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit a56587065094fd96eb4c2b5ad65571daad32156d)

MTK-Commit-Id: 87566f5eba6036a3edc9fecdbf86da77951807dd

Change-Id: I1efdf52e72f7625b562b0feac17dc5b68caa1dbc
Signed-off-by: Frankie Chang <Frankie.Chang@mediatek.com>
CR-Id: ALPS04825819
Feature: Google Security Patch