[ALPS05118439]vt: selection, push console lock down
We need to nest the console lock in sel_lock, so we have to push it down
a bit. Fortunately, the callers of set_selection_* just lock the console
lock around the function call. So moving it down is easy.

In the next patch, we switch the order.

MTK-Commit-Id: 835bf5e2562f3f80b9148038379f507432e635cb

Bug : 148713787
Feature: Google Security Patch
CR-Id: ALPS05118439
Change-Id: Id2c59634d86490a5771bfbe7d724a4b2f7544c4d
Signed-off-by: mtk21421 <Stan.Lu@mediatek.com>
(cherry-pick from e2b16b2bf15021456a72892171a92037d5db1afb)